completed

+++ entities/sprite.h
@@ -1,11 +1,23 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   sprite.h                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: apavel <apavel@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2020/11/04 13:16:30 by apavel            #+#    #+#             */
+/*   Updated: 2020/11/05 11:59:16 by apavel           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #ifndef SPRITE_H
 # define SPRITE_H
 
-typedef struct s_srpite
+typedef struct	s_srpite
 {
-	double x;
-	double y;
-	double distance;
+	float		x;
+	float		y;
+	float		distance;
 }				t_sprite;
 
 #endif
